+++ src/Home.jsx
@@ -29,7 +29,7 @@ export default function Home() {
   const navigate = useNavigate();
 
   React.useEffect(() => {
-    document.title = "MarketMindAI- AI Data Analytics";
+    document.title = "MArketMind-AI  FinOps for Marketing";
   }, []);
 
   return (